init: move timesync features to another file
